+++ pygeos/test/test_io.py
@@ -99,8 +99,17 @@ def test_from_wkb_exceptions():
     with pytest.raises(TypeError, match="Expected bytes, got int"):
         pygeos.from_wkb(1)
 
-    with pytest.raises(pygeos.GEOSException):
-        pygeos.from_wkb(b"\x01\x01\x00\x00\x00\x00")
+    # invalid WKB
+    with pytest.warns(Warning, match="Unexpected EOF parsing WKB"):
+        result = pygeos.from_wkb(b"\x01\x01\x00\x00\x00\x00")
+        assert result is None
+
+    # invalid ring in WKB
+    with pytest.warns(Warning, match="Invalid number of points in LinearRing found 3 - must be 0 or >= 4"):
+        result = pygeos.from_wkb(
+            b"\x01\x03\x00\x00\x00\x01\x00\x00\x00\x03\x00\x00\x00P}\xae\xc6\x00\xb15A\x00\xde\x02I\x8e^=A0n\xa3!\xfc\xb05A\xa0\x11\xa5=\x90^=AP}\xae\xc6\x00\xb15A\x00\xde\x02I\x8e^=A"
+        )
+        assert result is None
 
 
 @pytest.mark.parametrize("geom", all_types)